added login

+++ app.py
@@ -1,8 +1,10 @@
-from flask import Flask, redirect, render_template
+from flask import Flask, redirect, render_template,request
 from flask_sqlalchemy import SQLAlchemy
 from datetime import datetime
 from dbmethods import *
 from flask_migrate import Migrate
+from flask  import  session
+from sqlalchemy.orm import sessionmaker
 
 # TODO: fill db with items
 
@@ -16,6 +18,8 @@ app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///main.db'
 db = SQLAlchemy(app)
 # creating migration settings
 migrate = Migrate(app, db)
+# secret key
+app.secret_key = "dsagfhjsagkhjrgu123hjkgerkfhjsaghfjgh3qj1gwqruyaf"
 
 # model class for users
 class users(db.Model):
@@ -55,31 +59,34 @@ class basket(db.Model):
 # path for main index
 @app.route("/")
 def mainIndex():
-    # test current user
-    current_user = "admin"
-    # getting all items
-    allItems = items.query.all()
-    # TODO: create login
-    # getting all items in basket
-    itemsInBasket = basket.query.filter_by(owner=current_user).all()
-    # getting basket price
-    total = getTotal(itemsInBasket)
-    # rendering template
-    return render_template("index.html", items=allItems, basket=itemsInBasket,total=total)
+    if session.get("logged_in"):
+        # test current user
+        current_user = session.get("user")
+        # getting all items
+        allItems = items.query.all()
+        # TODO: create login
 
+        # getting all items in basket
+        itemsInBasket = basket.query.filter_by(owner=current_user).all()
+        # getting basket price
+
+        # doesnt work total
+        total = getTotal(itemsInBasket)
+
+        # TODO: filter items by cantegory
+
+        # rendering template
+        return render_template("index.html", items=allItems, basket=itemsInBasket,total=total)
+    else:
+        return redirect("/login")
 # тут будуть кнопочки для того шоб вибрати послугами
 # також буде корзина з вибраними послугами
 
 # path for adding item to basket
-# IDK IF IT WORKS CHECK IT
-# IDK IF IT WORKS CHECK IT
-# IDK IF IT WORKS CHECK IT
-# IDK IF IT WORKS CHECK IT
-
 @app.route("/addItemToBasket/<name>/<price>/<category>")
 def addItemToBasket(name, price, category):
     # test current user
-    current_user = "admin"
+    current_user = session.get("user")
     # checking if items exists
     if checkIfExists(name, price, category, items):
     # appending item into basket
@@ -114,23 +121,41 @@ def deleteitemFromItems(name, price, category):
 # path from deleting item from basket
 @app.route("/deleteitemfrombasket/<name>/<price>/<category>")
 def deleteItemFromBasket(name, price, category):
-    # getting item from basket
-    item = getItemFromBasket(name, price, category, basket, "admin") # current user
-    # deleting item from db
-    db.session.delete(item)
-    # saving changes of db
-    db.session.commit()
-    # redirecting into main index
-    return redirect("/")
+    if session.get("logged_ in"):
+        current_user = session.get("user")
+        # getting item from basket
+        item = getItemFromBasket(name, price, category, basket, current_user) # current user
+        # deleting item from db
+        db.session.delete(item)
+        # saving changes of db
+        db.session.commit()
+        # redirecting into main index
+        return redirect("/")
+    else:
+        return redirect("/login")
 
 @app.route("/checkout")
 def checkout():
-    current_user = "admin"
+    current_user = session.get("user")
     itemsInBasket = basket.query.filter_by(owner=current_user).all()
     deleteAllBasket(itemsInBasket, db)
     return redirect("/")
     # checkouting 
     # deleting items from basket
+
+@app.route("/login", methods=["POST", "GET"])
+def login():
+    session["logged_in"] = False
+    session["user"] = None
+    if request.method == "POST":
+        if checkAccess(request.form["password"], request.form["username"], users):
+            session["logged_in"] = True
+            session["user"] = request.form["username"]
+            return redirect("/")
+        else:
+            return redirect("/login")
+    else:
+        return render_template("login.html")
 
 if __name__ == "__main__":
     app.run(debug=True)
